requirements.txt
sseclient-py
kafka-python
pyspark==3.4.1
requests
dash==2.12.1
plotly==5.15.0
redis
pandas
